Changed loss type to SSIM3D

+++ models/w_net_3d.py
@@ -226,11 +226,11 @@ class WNet3D(nn.Module):
         encoder_op_mask = ip_mask * encoder_op
         class_prob = self.activation(encoder_op_mask)
         if ops == "enc":
-            return class_prob.float()
+            return class_prob
         reconstructed_op = self.Decoder(class_prob)
         if ops == "dec":
-            return reconstructed_op.float()
+            return reconstructed_op
         if ops == "both":
-            return class_prob.float(), reconstructed_op.float()
+            return class_prob, reconstructed_op
         else:
             raise ValueError('Invalid ops, ops must be in [enc, dec, both]')
